microservices-demo/release/authservice/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic==2.7.1
prometheus-fastapi-instrumentator==6.1.0
